WIP: new idea: make several gotos between two states
Instead of keeping only the "lowest" (in the ancestors graph) label
for gotos (e.g. NUM), keep them all (expr, term, NUM) as different
gotos.

The idea is that more gotos should allow for more precise lokbacks,
hence more precise lookaheads.

We also no longer track all the FOLLOWS of a goto from the outgoing
transitions of the state: it unions too many tokens, some of them
coming from other rules.  So rather, we go back to the kernel (with
closure) of the state, and look at the tokens that can follow the
non-terminal of the goto that we consider.